Crate gfx_window_glutin

Functions

init
Initialize with a window builder. Generically parametrized version over the main framebuffer format.

init_existing
Initialize with an existing Glutin window. Generically parametrized version over the main framebuffer format.

init_existing_raw
Initialize with an existing Glutin window. Raw version.

init_raw
Initialize with a window builder. Raw version.

new_views
Create new main target views based on the current size of the window. Best called just after a WindowResize event.

update_views
Update the internal dimensions of the main framebuffer targets. Generic version over the format.

update_views_raw
Return new main target views if the window resolution has changed from the old dimensions.